+++ extensions/night-light/src/turn-night-shift-off.ts
@@ -1,11 +1,9 @@
 import { nightlight } from "./utils";
-import { clearSearchBar, closeMainWindow, getPreferenceValues } from "@raycast/api";
+import { closeMainWindow, getPreferenceValues } from "@raycast/api";
 
 export default async function main() {
   const { closeWindow } = getPreferenceValues<{ closeWindow: boolean }>();
 
   if (closeWindow) await closeMainWindow();
-  else await clearSearchBar();
-
   await nightlight(["off"]);
 }